ADD Pressure to IAQ,  fix API variable and move iaq variable in air_quality .py
