+++ aioairzone_cloud/const.py
@@ -20,6 +20,7 @@ API_AQ_PM_10: Final[str] = "aqpm10"
 API_AQ_CO2: Final[str] = "aq_co2"
 API_AQ_TEMP: Final[str] = "aq_temp"
 API_AQ_TEMP_CELSIUS: Final[str] = "celsius"
+API_AQ_PRESSURE: Final[str] = "aq_pressure"
 API_AQ_HUMIDITY: Final[str] = "humidity"
 API_AQ_TVOC: Final[str] = "aq_tvoc"
 API_AQ_PRESENT: Final[str] = "aq_present"
@@ -29,6 +30,7 @@ API_AQ_QUALITY_LEVELS: Final[dict[str]] = {
     "regular": 151,
     "bad": 301,
 }
+API_AQ_VENT_ACTIVE: Final[str] = "aq_vent_active"
 API_AQ_SCORE: Final[str] = "aq_score"
 API_AQ_SENSOR_FW: Final[str] = "aq_sensor_fw"
 API_CELSIUS: Final[str] = "celsius"
@@ -172,6 +174,7 @@ AZD_AIR_DEMAND: Final[str] = "air-demand"
 AZD_AIR_QUALITY: Final[str] = "air-quality"
 AZD_AIR_QUALITY_ID: Final[str] = "air-quality-id"
 AZD_AQ_ACTIVE: Final[str] = "aq-active"
+AZD_AQ_VENT_ACTIVE: Final[str] = "aq-vent-active"
 AZD_AQ_SCORE: Final[str] = "aq-score"
 AZD_AQ_QUALITY: Final[str] = "aq-quality"
 AZD_AQ_MODE_CONF: Final[str] = "aq-mode-conf"
@@ -182,6 +185,7 @@ AZD_AQ_PM_10: Final[str] = "aq-pm-10"
 AZD_AQ_CO2: Final[str] = "aq-co2"
 AZD_AQ_TVOC: Final[str] = "aq-tvoc"
 AZD_AQ_TEMP: Final[str] = "aq-temperature"
+AZD_AQ_PRESSURE: Final[str] = "aq-pressure"
 AZD_AQ_HUMIDITY: Final[str] = "aq-humidity"
 AZD_AQ_PRESENT: Final[str] = "aq-present"
 AZD_AVAILABLE: Final[str] = "available"
